kmdc-menu - setAnchorCorder as Byte

+++ kmdc/kmdc-menu/src/jsMain/kotlin/MDCMenuModule.kt
@@ -2,8 +2,6 @@ package dev.petuska.kmdc.menu
 
 import dev.petuska.kmdc.core.Destroyable
 import dev.petuska.kmdc.core.MDCEvent
-import dev.petuska.kmdc.menu.surface.Corner
-import dev.petuska.kmdc.menu.surface.MDCMenuSurfaceModule
 import org.w3c.dom.Element
 import org.w3c.dom.HTMLElement
 
@@ -18,7 +16,7 @@ public external object MDCMenuModule {
 
     public var open: Boolean
 
-    public fun setAnchorCorner(corner: Int)
+    public fun setAnchorCorner(corner: Byte)
     public fun setAbsolutePosition(x: Double, y: Double)
     public fun setFixedPosition(isFixed: Boolean)
     public fun setSelectedIndex(index: Int)
